DEV: Define user-notes namespace earlier
Defining it inside `after_initialize` meant that other plugins couldn't always see it and interact with it during initalization.
